requirements.txt
Flask
openai==0.28
requests
beautifulsoup4
newspaper3k
python-dotenv
textblob
nltk
serpapi
lxml[html_clean]
pdfkit
google-search-results
fpdf
wkhtmltopdf